Make source const.
The compiler wasn't optimizing the one writeable field well. So now the
struct is fully constant and the sink data is stored in a function-local
static variable. This design is much cleaner as well as the source is
now always passed around as const rather then relying on a comment that
says "please only modify the sink_data field".